MovieJSONBuilder bug fix + package renaming
- Moved Grade API files into separate input_formatting and output_formatting packages.
- Fixed a change in the MovieJSONBuilder that prevented the cast, description, and runtime info from being converted to JSON.
- Overloaded the AbstractMedia constructor to provide a variation for users that already have a userRating set.
