+++ src/main/java/data_access/grade_api/incoming_data_formatting/MediaFactory.java
@@ -1,4 +1,4 @@
-package data_access.grade_api;
+package data_access.grade_api.incoming_data_formatting;
 
 import java.util.List;
